refactored nav folder
