+++ hamza-client/src/modules/layout/templates/nav-4/connect-button/connect-button.tsx
@@ -2,7 +2,7 @@
 
 import { Button, Flex } from '@chakra-ui/react';
 import { ConnectButton } from '@rainbow-me/rainbowkit';
-import AccountMenu from '@modules/layout/templates/nav-4/components/menu/account-menu';
+import AccountMenu from '@modules/layout/templates/nav-4/menu/account-menu';
 import CartButton from '../cart-button/components/cart-button';
 
 //Todo: If chain unsupported?
